Fix applying video settings.
If applying settings after pressing the back key in the video options
menu after changing settings, the Yes and No options would have the
opposite effect, with No applying the changes and Yes reverting them.

I've also pointed out a bug in the code, which causes the settings to
not be saved when the changes are applied. I can confirm that this
bug exists in the latest Steam EXE.