Fix shader option setting 'changes pending' flag.
In vanilla Mania, changing the shader setting and then leaving the
menu by pressing the back button does *not* cause the 'apply changes'
prompt to appear.

In this decompilation, the prompt would appear, but only if you
backed out of the menu, re-entered it, and backed out again. This
behaviour does not occur in the latest Steam EXE.