Fix Pinball bugs
This fixes two bugs:

- If you get multiple different shields in the crane game, then the
character may have a bugged 'null' shield upon returning to the
level. This is caused by the IDs of the obtained shields being OR'd
together, producing an invalid ID.

- In Mania Mode, the crane game should offer 1UPs, but an incorrect
`RSDK::Rand` parameter was preventing it from doing so.

Both of these fixes have been verified against the assembly code of
the latest Steam EXE.