Fix Insta-Shield hitbox
The `tempHitbox` variable was only valid inside of the if, which
lead to undefined behavior when using its address afterwards.
While not an issue on MSVC, more aggressive compilers like GCC (O3)
and Clang (O1) took advantage of this range limit to further
optimize the code, and caused all sorts of bugs on Linux/Android.