CutsceneSeq: Split CutsceneSeq_SetSkipType function
In the original code, cutscene objects do not set the skipCallback field
unless when using SKIPTYPE_CALLBACK.
Make CutsceneSeq_SetSkipType only set the skip type (for non-callback skips),
and introduce a new CutsceneSeq_SetSkipTypeCallback which sets the skip type
as SKIPTYPE_CALLBACK and only takes the callback function as parameter.
Also add them in the public function list so they can be used in mods.