Fix incorrect anim order in PlayerAnimationIDs
Fixes `ANI_POLE_SWING_H` and `ANI_BREATHE` going before `ANI_POLE_SWING_V` and `ANI_BUBBLE` respectively, both of which not being the case in the animation files. This made the sections of the code these animations are used in difficult to understand due to the wrong animation names being listed.