CMake: Disable COMDAT folding on MSVC
State checks can break on MSVC because Player_State_Static and
Player_State_TransportTube (maybe more) are empty and are therefore
identical and can be merged.
While GCC/Clang knows not to merge both since their addresses are being
referenced in the code, but MSVC thinks it's a smart optimization to do...

Fixes #303.